Rename utils -> definitions
