+++ lib/i18next.dart
@@ -2,8 +2,8 @@ library i18next;
 
 export 'src/data_sources/asset_bundle_data_source.dart';
 export 'src/data_sources/localization_data_source.dart';
+export 'src/definitions.dart';
 export 'src/i18next.dart';
 export 'src/i18next_localization_delegate.dart';
 export 'src/options.dart';
 export 'src/resource_store.dart';
-export 'src/utils.dart';
